refactor: move get event participants API from UserController to EventController
