+++ src/main/java/lems/cowshed/api/controller/regular/event/RegularEventController.java
@@ -2,7 +2,7 @@ package lems.cowshed.api.controller.regular.event;
 
 import jakarta.validation.Valid;
 import lems.cowshed.api.controller.CommonResponse;
-import lems.cowshed.api.controller.dto.regular.event.RegularEventSaveRequest;
+import lems.cowshed.api.controller.dto.regular.event.request.RegularEventSaveRequest;
 import lems.cowshed.domain.user.CustomUserDetails;
 import lems.cowshed.service.RegularEventService;
 import lombok.RequiredArgsConstructor;
