+++ src/test/java/lems/cowshed/domain/event/query/EventQueryRepositoryTest.java
@@ -2,6 +2,7 @@ package lems.cowshed.domain.event.query;
 
 import lems.cowshed.IntegrationTestSupport;
 import lems.cowshed.api.controller.dto.event.response.EventSimpleInfo;
+import lems.cowshed.api.controller.dto.event.response.query.EventParticipantQueryDto;
 import lems.cowshed.domain.bookmark.Bookmark;
 import lems.cowshed.domain.bookmark.BookmarkRepository;
 import lems.cowshed.domain.event.Event;
@@ -192,6 +193,66 @@ class EventQueryRepositoryTest extends IntegrationTestSupport {
         assertThat(result).isNotEmpty()
                 .extracting("name", "content", "bookmarkStatus")
                 .containsExactly(Tuple.tuple("산책 모임", "테스트 내용", BOOKMARK));
+    }
+
+    @DisplayName("모임에 참여한 회원들을 조회 한다.")
+    @Test
+    void getEventParticipants() {
+        //given
+        User user = createUser("테스터", INTP);
+        userRepository.save(user);
+        Event event = createEvent("산책 모임", "테스터");
+        eventRepository.save(event);
+        EventParticipation eventParticipation = EventParticipation.of(user, event);
+        eventParticipantRepository.save(eventParticipation);
+
+        //when
+        List<EventParticipantQueryDto> result = eventQueryRepository.getEventParticipants(event.getId());
+
+        //then
+        assertThat(result.get(0))
+                .extracting("name", "mbti")
+                .containsExactly("테스터", INTP);
+    }
+
+    @DisplayName("모임에 참여한 회원들을 조회 할때 참여한 회원이 없다면 빈 리스트를 반환 한다.")
+    @Test
+    void getEventParticipants_WhenZeroParticipants_ThenReturnEmptyList() {
+        //given
+        User user = createUser("테스터", INTP);
+        userRepository.save(user);
+        Event event = createEvent("산책 모임", "테스터");
+        eventRepository.save(event);
+
+        //when
+        List<EventParticipantQueryDto> result = eventQueryRepository.getEventParticipants(event.getId());
+
+        //then
+        assertThat(result).isEmpty();
+    }
+
+    @DisplayName("모임에 참여한 회원들을 조회 할때 두명의 회원이 참여한 경우 모임의 참여 인원수는 2명 이다.")
+    @Test
+    void getEventParticipants_WhenTwoUsersParticipate_ThenCountIsTwo() {
+        //given
+        User user = createUser("테스터", INTP);
+        User user2 = createUser("테스터2", ESFJ);
+        Event event = createEvent("산책 모임", "테스터");
+
+        EventParticipation eventParticipation = EventParticipation.of(user, event);
+        EventParticipation eventParticipation2 = EventParticipation.of(user2, event);
+
+        userRepository.saveAll(List.of(user, user2));
+        eventRepository.save(event);
+        eventParticipantRepository.saveAll(List.of(eventParticipation, eventParticipation2));
+
+        //when
+        List<EventParticipantQueryDto> result = eventQueryRepository.getEventParticipants(event.getId());
+
+        //then
+        assertThat(result).hasSize(2)
+                .extracting("name")
+                .containsExactlyInAnyOrder("테스터", "테스터2");
     }
 
     private RegularEvent createRegularEvent(Event event, String name, String location){
